Move AuthHandshakeMessageHandler to new file
